+++ minigalaxy/ui/library.py
@@ -1,5 +1,4 @@
 import json
-import locale
 import logging
 import os
 import re
@@ -33,16 +32,6 @@ class Library(Gtk.Viewport):
 
         self.parent_window = parent_window
         self.config = config
-
-        current_locale = self.config.locale
-        default_locale = locale.getlocale()[0]
-        if current_locale == '':
-            locale.setlocale(locale.LC_ALL, (default_locale, 'UTF-8'))
-        else:
-            try:
-                locale.setlocale(locale.LC_ALL, (current_locale, 'UTF-8'))
-            except NameError:
-                locale.setlocale(locale.LC_ALL, (default_locale, 'UTF-8'))
 
         self.api = api
         self.download_manager = download_manager